Move bool parameter to parameter module
